+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Dodo x.x.x (not yet released)
+* Update to IDLeDOM 0.7.1.
 * Fix doctype creation and "significant whitespace" handling in DOMParser.
 * Ensure that Document::documentElement is always populated.
-* Update to IDLeDOM 0.7.1.
+* EXPERIMENTAL: Added Node::getExtensionData() and Node::setExtensionData()
+  methods to allow end-users to associate additional non-spec data off of
+  Nodes.
 
 # Dodo 0.1.0 (2021-07-04)
 Initial release.